+++ src/main/java/com/example/travelday/domain/travelplan/dto/request/group/UpdateTravelPlansReqDto.java
@@ -1,4 +1,4 @@
-package com.example.travelday.domain.travelplan.dto.request.batch;
+package com.example.travelday.domain.travelplan.dto.request.group;
 
 import com.example.travelday.domain.travelplan.dto.request.TravelPlanOverwriteReqDto;
 import jakarta.validation.constraints.NotEmpty;
@@ -7,7 +7,7 @@ import lombok.Builder;
 import java.util.List;
 
 @Builder
-public record UpdateTravelPlanListReqDto(
+public record UpdateTravelPlansReqDto(
 
     @NotEmpty
     List<TravelPlanOverwriteReqDto> body
